comm: refactor and initialize `flashinfer.comm` module (#1089)
