+++ .gitmodules
@@ -4,9 +4,6 @@
 [submodule "3rdparty/googletest"]
 	path = 3rdparty/googletest
 	url = https://github.com/google/googletest.git
-[submodule "3rdparty/mscclpp"]
-	path = 3rdparty/mscclpp
-	url = https://github.com/microsoft/mscclpp.git
 [submodule "3rdparty/cutlass"]
 	path = 3rdparty/cutlass
 	url = https://github.com/NVIDIA/cutlass.git
